backend/requirements.txt
langchain
langchain-core
langchain-text-splitters
langgraph
langsmith
pydantic
python-dotenv
fastapi
uvicorn
langchain-openai
langchain-google-genai
langchain-groq
langchain-anthropic
langchain-ollama
markitdown
python-multipart
python-docx
supabase
redis
rapidfuzz
tiktoken
cryptography
requests
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytest
httpx
